Build "native" gems for windows that don't try to invoke make
Since protoc won't ever be re-built on Windows, there's no need to
include the "extension" that build it on those platforms.  There is also
a fairly high likelyhood that no 'make' command exists on Windows, so
having a version of the gem that doesn't try to invoke make will make
this gem easier to use on Windows.